refactor(portal): samle ColorScheme-definisjon i temabyggeren
