+++ .changeset/curly-rivers-spark.md
@@ -1,0 +1,5 @@
+---
+"portal": patch
+---
+
+Samler definisjonen av fargetemaer i temabyggeren, slik at `ColorScheme` og `COLOR_SCHEMES` har samme kilde.
